Make `PriceLoader` a reusable component
